% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_core_shared_hotspots.R
\name{analyze_core_shared_hotspots}
\alias{analyze_core_shared_hotspots}
\title{Analyze core shared hotspots}
\usage{
analyze_core_shared_hotspots(candidate_hotspots, config, task_params = NULL)
}
\arguments{
\item{candidate_hotspots}{Data frame of candidate hotspots}

\item{config}{Configuration object}

\item{task_params}{Task parameters}
}
\value{
List containing core shared genes analysis results
}
\description{
Analyze core shared hotspots
}
