% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_mapping.R
\name{apply_column_mapping}
\alias{apply_column_mapping}
\title{Apply column mapping to data frame}
\usage{
apply_column_mapping(data, mapping_list, file_type = "main_data")
}
\arguments{
\item{data}{Original data frame}

\item{mapping_list}{List of column mappings (standard_name -> original_name)}

\item{file_type}{Type of file for validation}
}
\value{
Data frame with renamed columns
}
\description{
Apply column mapping to data frame
}
