% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_parser.R
\name{apply_execution_mode_to_tasks}
\alias{apply_execution_mode_to_tasks}
\title{Apply execution mode adjustments to tasks configuration}
\usage{
apply_execution_mode_to_tasks(tasks_config, mode)
}
\arguments{
\item{tasks_config}{Analysis tasks configuration}

\item{mode}{Execution mode}
}
\value{
Adjusted tasks configuration
}
\description{
Apply execution mode adjustments to tasks configuration
}
