% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_parser.R
\name{apply_mode_to_config}
\alias{apply_mode_to_config}
\title{Apply execution mode to configuration}
\usage{
apply_mode_to_config(config_data, mode)
}
\arguments{
\item{config_data}{Configuration data}

\item{mode}{Execution mode}
}
\value{
Adjusted configuration data
}
\description{
Apply execution mode to configuration
}
