% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_species_ordering.R
\name{apply_species_ordering}
\alias{apply_species_ordering}
\title{Reorder a species factor based on a provided order vector (Refactored for Compliance)
Rule #6 & #9 Compliant: No longer reads files or uses config.}
\usage{
apply_species_ordering(data, species_order_vector, species_col = "species")
}
\arguments{
\item{data}{Data frame containing species column}

\item{species_order_vector}{A character vector defining the desired order of species.}

\item{species_col}{The name of the species column in the data frame.}
}
\value{
Data frame with the species column as an ordered factor.
apply_species_ordering
}
\description{
Reorder a species factor based on a provided order vector (Refactored for Compliance)
Rule #6 & #9 Compliant: No longer reads files or uses config.
}
