% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_parser.R
\name{apply_visualization_only_adjustments}
\alias{apply_visualization_only_adjustments}
\title{Apply visualization-only mode adjustments to a task}
\usage{
apply_visualization_only_adjustments(task, task_name)
}
\arguments{
\item{task}{Task configuration}

\item{task_name}{Task name}
}
\value{
Adjusted task configuration
}
\description{
Apply visualization-only mode adjustments to a task
}
