% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_correlation_pvalues.R
\name{calculate_correlation_pvalues}
\alias{calculate_correlation_pvalues}
\title{Calculate correlation p-values for data matrix}
\usage{
calculate_correlation_pvalues(data_matrix, method = "spearman")
}
\arguments{
\item{data_matrix}{Numeric matrix for correlation analysis}

\item{method}{Correlation method to use (default: "spearman")}
}
\value{
Matrix of p-values for pairwise correlations
}
\description{
Calculates pairwise correlation p-values between rows of a data matrix
}
\details{
Calculate correlation p-values for data matrix
}
