% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_layout_utils.R
\name{calculate_grid_dimensions}
\alias{calculate_grid_dimensions}
\title{Calculate Optimal Grid Dimensions}
\usage{
calculate_grid_dimensions(n_plots, max_cols = 3)
}
\arguments{
\item{n_plots}{Number of plots to arrange}

\item{max_cols}{Maximum number of columns allowed}
}
\value{
List with nrow and ncol values
}
\description{
Helper function to calculate the optimal number of rows and columns
for arranging a given number of plots with a maximum column constraint.
}
