% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_processing.R
\name{calculate_igs_regions}
\alias{calculate_igs_regions}
\title{Calculate IGS (intergenic sequences) using GenomicRanges::gaps()}
\usage{
calculate_igs_regions(gff_object, species_name)
}
\arguments{
\item{gff_object}{Validated GRanges object}

\item{species_name}{Species name}
}
\value{
Data frame with IGS regions
}
\description{
Calculate IGS (intergenic sequences) using GenomicRanges::gaps()
}
