% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M01_helpers_sequence_variants.R
\name{calculate_igs_snp_indel_ratio}
\alias{calculate_igs_snp_indel_ratio}
\title{Calculate IGS SNP/INDEL Ratio}
\usage{
calculate_igs_snp_indel_ratio(filtered_data_file, pseudocount)
}
\arguments{
\item{filtered_data_file}{Path to variants_filtered.csv}

\item{pseudocount}{Pseudocount value}
}
\value{
Data frame with SNP/INDEL ratios
}
\description{
Calculate IGS SNP/INDEL Ratio
}
\keyword{internal}
