% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M01_helpers_sequence_variants.R
\name{calculate_indel_length_stats}
\alias{calculate_indel_length_stats}
\title{Calculate INDEL Length Statistics}
\usage{
calculate_indel_length_stats(indel_data, min_indel_length)
}
\arguments{
\item{indel_data}{INDEL data frame}

\item{min_indel_length}{Minimum length threshold}
}
\value{
List with statistics
}
\description{
Calculate INDEL Length Statistics
}
\keyword{internal}
