% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_processing.R
\name{calculate_intron_regions}
\alias{calculate_intron_regions}
\title{Calculate Intron regions using GenomicRanges::gaps() within genes}
\usage{
calculate_intron_regions(gff_object, species_name)
}
\arguments{
\item{gff_object}{Validated GRanges object}

\item{species_name}{Species name}
}
\value{
Data frame with Intron regions
}
\description{
Calculate Intron regions using GenomicRanges::gaps() within genes
}
