% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_final_report.R
\name{calculate_quality_scores}
\alias{calculate_quality_scores}
\title{Calculate quality assessment scores}
\usage{
calculate_quality_scores(stats)
}
\arguments{
\item{stats}{Statistics object}
}
\value{
Quality scores list
}
\description{
Calculate quality assessment scores
}
