% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M01_helpers_sequence_variants.R
\name{calculate_sn_ratio}
\alias{calculate_sn_ratio}
\title{Calculate S/N Ratio}
\usage{
calculate_sn_ratio(site_classification, pseudocount, min_sites_per_gene)
}
\arguments{
\item{site_classification}{Data frame with site types}

\item{pseudocount}{Pseudocount for ratio calculation}

\item{min_sites_per_gene}{Minimum sites per gene}
}
\value{
Data frame with S/N ratios
}
\description{
Calculate S/N Ratio
}
\keyword{internal}
