% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M01_helpers_sequence_variants.R
\name{calculate_snp_indel_ratio_by_genome_region}
\alias{calculate_snp_indel_ratio_by_genome_region}
\title{Calculate SNP/INDEL Ratio by Genome Region}
\usage{
calculate_snp_indel_ratio_by_genome_region(filtered_data_file, pseudocount)
}
\arguments{
\item{filtered_data_file}{Path to P02 filtered data}

\item{pseudocount}{Pseudocount for ratio calculation}
}
\value{
Data frame with region-level SNP/INDEL statistics
}
\description{
Calculate SNP/INDEL ratio for all region types grouped by genome region (LSC/SSC/IRA)
}
\keyword{internal}
