% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_validate_region_info_v2.R
\name{check_gene_names_v2}
\alias{check_gene_names_v2}
\title{Check Gene Names V2}
\usage{
check_gene_names_v2(raw_data, gene_function_map_path = NULL)
}
\description{
Check Gene Names V2
}
\keyword{internal}
