% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_validate_region_info_v2.R
\name{check_negative_coordinates_v2}
\alias{check_negative_coordinates_v2}
\title{Check Negative Coordinates V2}
\usage{
check_negative_coordinates_v2(raw_data)
}
\description{
Check Negative Coordinates V2
}
\keyword{internal}
