% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{clean_species_names}
\alias{clean_species_names}
\title{Clean and standardize species names}
\usage{
clean_species_names(species_names)
}
\arguments{
\item{species_names}{Character vector of species names to clean}
}
\description{
Clean and standardize species names
}
