% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_utils.R
\name{cleanup_memory}
\alias{cleanup_memory}
\title{Memory cleanup function}
\usage{
cleanup_memory(verbose = FALSE)
}
\arguments{
\item{verbose}{Logical, whether to print verbose output (default: FALSE)}
}
\value{
Invisible garbage collection results
}
\description{
Performs garbage collection to free up memory
}
\details{
Memory cleanup function
}
