% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M01_helpers_sequence_variants.R
\name{cluster_species_by_shared_indels}
\alias{cluster_species_by_shared_indels}
\title{Cluster Species by Shared INDELs (Jaccard Similarity)}
\usage{
cluster_species_by_shared_indels(upset_matrix)
}
\arguments{
\item{upset_matrix}{Binary matrix of sequences x species}
}
\value{
List with hclust object and distance matrix
}
\description{
Cluster Species by Shared INDELs (Jaccard Similarity)
}
\keyword{internal}
