% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_final_report.R
\name{collect_data_quality_summary}
\alias{collect_data_quality_summary}
\title{Collect data quality summary}
\usage{
collect_data_quality_summary(session_paths)
}
\arguments{
\item{session_paths}{Session paths object}
}
\value{
Data quality summary string
}
\description{
Collect data quality summary
}
