% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_final_report.R
\name{collect_enabled_modules}
\alias{collect_enabled_modules}
\title{Collect enabled modules from configuration}
\usage{
collect_enabled_modules(config_data)
}
\arguments{
\item{config_data}{Configuration data}
}
\value{
String of enabled modules
}
\description{
Collect enabled modules from configuration
}
