% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_final_report.R
\name{collect_enhanced_enabled_modules}
\alias{collect_enhanced_enabled_modules}
\title{Collect enhanced enabled modules information}
\usage{
collect_enhanced_enabled_modules(config_data)
}
\arguments{
\item{config_data}{Configuration data}
}
\value{
Enhanced enabled modules information
}
\description{
Collect enhanced enabled modules information
}
