% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_final_report.R
\name{collect_file_statistics}
\alias{collect_file_statistics}
\title{Collect file statistics}
\usage{
collect_file_statistics(session_paths)
}
\arguments{
\item{session_paths}{Session paths object}
}
\value{
List of file statistics
}
\description{
Collect file statistics
}
