% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_final_report.R
\name{collect_module_files}
\alias{collect_module_files}
\title{Collect and organize files by module}
\usage{
collect_module_files(session_paths)
}
\arguments{
\item{session_paths}{Session paths object}
}
\value{
List of file collections by module
}
\description{
Collect and organize files by module
}
