% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_final_report.R
\name{collect_resource_usage_info}
\alias{collect_resource_usage_info}
\title{Collect resource usage information}
\usage{
collect_resource_usage_info(session_paths)
}
\arguments{
\item{session_paths}{Session paths object}
}
\value{
Resource usage information
}
\description{
Collect resource usage information
}
