% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_dispatcher.R
\name{collect_task_manifests}
\alias{collect_task_manifests}
\title{Collect manifests from all task results}
\usage{
collect_task_manifests(execution_context)
}
\arguments{
\item{execution_context}{Execution context containing task results}
}
\value{
List of manifests from all successful tasks
}
\description{
Collect manifests from all task results
}
