% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_final_report.R
\name{collect_timing_information}
\alias{collect_timing_information}
\title{Collect enhanced timing information}
\usage{
collect_timing_information(execution_context, session_paths)
}
\arguments{
\item{execution_context}{Execution context}

\item{session_paths}{Session paths object}
}
\value{
Timing information list
}
\description{
Collect enhanced timing information
}
