% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_frequencies.R
\name{combine_frequency_data}
\alias{combine_frequency_data}
\title{Combine frequency data from different regions}
\usage{
combine_frequency_data(
  cds_frequencies,
  igs_frequencies,
  intron_frequencies,
  other_frequencies = data.frame()
)
}
\arguments{
\item{cds_frequencies}{CDS frequency data}

\item{igs_frequencies}{IGS frequency data}

\item{intron_frequencies}{Intron frequency data}

\item{other_frequencies}{Additional frequency data from other regions}
}
\value{
Combined frequency data with standardized structure
}
\description{
Combine frequency data from different regions
}
