% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_validator.R
\name{convert_nested_list_to_df}
\alias{convert_nested_list_to_df}
\title{Convert nested list to data frame format}
\usage{
convert_nested_list_to_df(nested_list)
}
\arguments{
\item{nested_list}{SPECIES_GENOME_REGIONS format list}
}
\value{
Data frame representation
}
\description{
Convert nested list to data frame format
}
