% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_converter.R
\name{convert_pdf_to_jpg}
\alias{convert_pdf_to_jpg}
\title{Convert PDF to JPG}
\usage{
convert_pdf_to_jpg(pdf_path, jpg_path, config)
}
\arguments{
\item{pdf_path}{Input PDF file path}

\item{jpg_path}{Output JPG file path}

\item{config}{Visualization configuration}
}
\value{
TRUE if successful
}
\description{
Convert PDF to JPG
}
