% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_converter.R
\name{convert_pdf_to_png}
\alias{convert_pdf_to_png}
\title{Convert PDF to PNG}
\usage{
convert_pdf_to_png(pdf_path, png_path, config)
}
\arguments{
\item{pdf_path}{Input PDF file path}

\item{png_path}{Output PNG file path}

\item{config}{Visualization configuration}
}
\value{
TRUE if successful
}
\description{
Convert PDF to PNG
}
