% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_converter.R
\name{convert_pdf_to_svg}
\alias{convert_pdf_to_svg}
\title{Convert PDF to SVG}
\usage{
convert_pdf_to_svg(pdf_path, svg_path, config)
}
\arguments{
\item{pdf_path}{Input PDF file path}

\item{svg_path}{Output SVG file path}

\item{config}{Visualization configuration}
}
\value{
TRUE if successful
}
\description{
Convert PDF to SVG
}
