% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_final_report.R
\name{convert_to_relative_path}
\alias{convert_to_relative_path}
\title{Convert absolute path to relative path for report (V6.5 Fix)}
\usage{
convert_to_relative_path(absolute_path, report_base_path)
}
\arguments{
\item{absolute_path}{Full path to file}

\item{report_base_path}{Base path where report will be located}
}
\value{
Relative path for use in href attributes
}
\description{
Convert absolute path to relative path for report (V6.5 Fix)
}
