% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_annotated_correlation_text.R
\name{create_annotated_correlation_text}
\alias{create_annotated_correlation_text}
\title{Create annotated correlation text matrix}
\usage{
create_annotated_correlation_text(cor_matrix, pvalue_matrix)
}
\arguments{
\item{cor_matrix}{Matrix of correlation coefficients}

\item{pvalue_matrix}{Matrix of p-values corresponding to correlations}
}
\value{
Matrix of formatted text combining correlations and significance
}
\description{
Creates text annotations combining correlation values with significance stars
}
\details{
Create annotated correlation text matrix
}
