% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_layout_utils.R
\name{create_arrangement_summary}
\alias{create_arrangement_summary}
\title{Create Plot Grid Summary}
\usage{
create_arrangement_summary(plot_names, grid_dims)
}
\arguments{
\item{plot_names}{Names of the plots being arranged}

\item{grid_dims}{Grid dimensions (from calculate_grid_dimensions)}
}
\value{
Summary ggplot object
}
\description{
Creates a summary information plot showing the arrangement details
}
