% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_module_utils.R
\name{create_data_summary}
\alias{create_data_summary}
\title{Create data summary information}
\usage{
create_data_summary(data)
}
\arguments{
\item{data}{Data frame}
}
\value{
String containing data summary
}
\description{
Create data summary information
}
