% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_utils.R
\name{create_dir}
\alias{create_dir}
\title{Create directory if it doesn't exist}
\usage{
create_dir(path, recursive = TRUE)
}
\arguments{
\item{path}{Directory path to create}

\item{recursive}{Whether to create parent directories if needed}
}
\value{
TRUE on success
}
\description{
Create directory if it doesn't exist
}
