% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_helpers.R
\name{create_explanatory_subtitle}
\alias{create_explanatory_subtitle}
\title{Create explanatory subtitle for different plot types (Refactored for Compliance)
Rule #6 Compliant: No longer takes config object. Receives a simple list of parameters.}
\usage{
create_explanatory_subtitle(plot_type, params_list = list())
}
\arguments{
\item{plot_type}{Type of plot for which to create subtitle}

\item{params_list}{List of parameters for subtitle generation}
}
\value{
Character string with explanatory subtitle
}
\description{
Create explanatory subtitle for different plot types (Refactored for Compliance)
Rule #6 Compliant: No longer takes config object. Receives a simple list of parameters.
}
