% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_utils.R
\name{create_filename}
\alias{create_filename}
\title{Create standardized file names}
\usage{
create_filename(
  base_name,
  suffix = NULL,
  extension = "csv",
  add_timestamp = FALSE
)
}
\arguments{
\item{base_name}{Base name for the file}

\item{suffix}{Optional suffix to add to filename}

\item{extension}{File extension}

\item{add_timestamp}{Whether to add timestamp to filename}
}
\value{
Standardized filename string
}
\description{
Create standardized file names
}
