% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest_utils.R
\name{create_manifest_collection}
\alias{create_manifest_collection}
\title{Create a manifest collection for a module}
\usage{
create_manifest_collection(entries, module_name, session_id)
}
\arguments{
\item{entries}{List of manifest entries}

\item{module_name}{Name of the module (M01, M02, M03)}

\item{session_id}{Session identifier}
}
\value{
Manifest collection (list)
}
\description{
Create a manifest collection for a module
}
