% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest_utils.R
\name{create_manifest_entry}
\alias{create_manifest_entry}
\title{Create a standardized output manifest entry}
\usage{
create_manifest_entry(
  path,
  role,
  title,
  caption,
  module = NULL,
  file_type = NULL,
  priority = 5,
  parameters = NULL
)
}
\arguments{
\item{path}{Absolute path to the output file}

\item{role}{Role of the file (composite_plot, individual_plot, summary_table, raw_data, etc.)}

\item{title}{Academic title for the output (e.g., "Figure 1: Distribution Analysis")}

\item{caption}{Detailed caption/description of the output}

\item{module}{Source module that generated this output (M01, M02, M03, etc.)}

\item{file_type}{Type of file (plot, table, data, report)}

\item{priority}{Display priority (1=highest, used for report ordering)}

\item{parameters}{Analysis parameters used for this output (V7.1 Enhancement)}
}
\value{
Standardized manifest entry (list)
create_manifest_entry
}
\description{
Create a standardized output manifest entry
}
