% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_module_pdfs.R
\name{create_pdf_manifest}
\alias{create_pdf_manifest}
\title{Create a manifest of all PDF files in the session}
\usage{
create_pdf_manifest(session_path)
}
\arguments{
\item{session_path}{Character. Path to the session directory}
}
\value{
Data frame. Manifest of all PDF files with metadata
}
\description{
Create a manifest of all PDF files in the session
}
