% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_variants.R
\name{create_preprocessing_summary}
\alias{create_preprocessing_summary}
\title{Create preprocessing summary statistics}
\usage{
create_preprocessing_summary(processed_data, original_data)
}
\arguments{
\item{processed_data}{Final processed data}

\item{original_data}{Original raw data}
}
\value{
Summary statistics data frame
}
\description{
Create preprocessing summary statistics
}
