% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_mapping.R
\name{create_sample_data}
\alias{create_sample_data}
\title{Create sample data for testing different file types}
\usage{
create_sample_data(n_rows = 100, file_type = "main_data")
}
\arguments{
\item{n_rows}{Number of rows to generate}

\item{file_type}{Type of file to generate}
}
\value{
Data frame with sample data
}
\description{
Create sample data for testing different file types
}
