% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_parser.R
\name{create_sequential_execution_plan}
\alias{create_sequential_execution_plan}
\title{Create sequential execution plan with dependency resolution}
\usage{
create_sequential_execution_plan(tasks)
}
\arguments{
\item{tasks}{List of enabled tasks}
}
\value{
Sequential execution plan
}
\description{
Create sequential execution plan with dependency resolution
}
