% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_parser.R
\name{create_task_execution_plan}
\alias{create_task_execution_plan}
\title{Create task execution plan}
\usage{
create_task_execution_plan(tasks, execution_config)
}
\arguments{
\item{tasks}{List of parsed tasks}

\item{execution_config}{Task execution configuration}
}
\value{
Execution plan with ordered task stages
}
\description{
Create task execution plan
}
