% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_parser.R
\name{estimate_task_duration}
\alias{estimate_task_duration}
\title{Estimate task execution duration}
\usage{
estimate_task_duration(task)
}
\arguments{
\item{task}{Task object}
}
\value{
Estimated duration in minutes
}
\description{
Estimate task execution duration
}
