% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_final_report.R
\name{estimate_timing_from_files}
\alias{estimate_timing_from_files}
\title{Estimate timing from file timestamps}
\usage{
estimate_timing_from_files(session_paths)
}
\arguments{
\item{session_paths}{Session paths object}
}
\value{
Estimated timing information
}
\description{
Estimate timing from file timestamps
}
