% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_validator.R
\name{evaluate_simple_condition}
\alias{evaluate_simple_condition}
\title{Simple condition evaluator (basic implementation)}
\usage{
evaluate_simple_condition(row_data, condition)
}
\arguments{
\item{row_data}{Single row of configuration data}

\item{condition}{String condition to evaluate}
}
\value{
Logical result
}
\description{
Simple condition evaluator (basic implementation)
}
