% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_dispatcher.R
\name{execute_filtering_task}
\alias{execute_filtering_task}
\title{Execute filtering task}
\usage{
execute_filtering_task(task_config, execution_context, parameters)
}
\arguments{
\item{task_config}{Task configuration}

\item{execution_context}{Execution context}

\item{parameters}{Merged parameters}
}
\value{
Filtering result
}
\description{
Execute filtering task
}
