% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_dispatcher.R
\name{execute_generate_report_task}
\alias{execute_generate_report_task}
\title{Execute report generation task}
\usage{
execute_generate_report_task(task_config, execution_context, parameters)
}
\arguments{
\item{task_config}{Task configuration}

\item{execution_context}{Execution context}

\item{parameters}{Merged parameters}
}
\value{
Report generation result
}
\description{
Execute report generation task
}
