% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_dispatcher.R
\name{execute_M02_task}
\alias{execute_M02_task}
\title{Execute M02 comparative analysis task}
\usage{
execute_M02_task(task_config, execution_context, parameters)
}
\arguments{
\item{task_config}{Task configuration}

\item{execution_context}{Execution context}

\item{parameters}{Merged parameters}
}
\value{
M02 result
}
\description{
Execute M02 comparative analysis task
}
