% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_dispatcher.R
\name{execute_normalization_task}
\alias{execute_normalization_task}
\title{Execute normalization task}
\usage{
execute_normalization_task(task_config, execution_context, parameters)
}
\arguments{
\item{task_config}{Task configuration}

\item{execution_context}{Execution context}

\item{parameters}{Merged parameters}
}
\value{
Normalization result
}
\description{
Execute normalization task
}
