% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_dispatcher.R
\name{execute_preprocessing_task}
\alias{execute_preprocessing_task}
\title{Execute preprocessing task}
\usage{
execute_preprocessing_task(task_config, execution_context, parameters)
}
\arguments{
\item{task_config}{Task configuration}

\item{execution_context}{Execution context}

\item{parameters}{Merged parameters}
}
\value{
Preprocessing result
}
\description{
Execute preprocessing task
}
