% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_yaml_utils.R
\name{export_yaml_config}
\alias{export_yaml_config}
\title{Export configuration to YAML file}
\usage{
export_yaml_config(config_data, output_file)
}
\arguments{
\item{config_data}{Configuration data list}

\item{output_file}{Path for output YAML file}
}
\value{
TRUE if successful
}
\description{
Export configuration to YAML file
}
