% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_processing.R
\name{extract_cds_features}
\alias{extract_cds_features}
\title{Extract CDS features from GFF object}
\usage{
extract_cds_features(gff_object, species_name)
}
\arguments{
\item{gff_object}{Validated GRanges object}

\item{species_name}{Species name}
}
\value{
Data frame with CDS features
}
\description{
Extract CDS features from GFF object
}
