% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M01_helpers_sequence_variants.R
\name{extract_filtered_cds_snp_positions}
\alias{extract_filtered_cds_snp_positions}
\title{Extract Filtered CDS SNP Positions}
\usage{
extract_filtered_cds_snp_positions(filtered_data_file)
}
\arguments{
\item{filtered_data_file}{Path to P02_filtered/variants_filtered.csv}
}
\value{
Data frame with species, position
}
\description{
Extract Filtered CDS SNP Positions
}
\keyword{internal}
