% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M01_helpers_sequence_variants.R
\name{extract_filtered_igs_indel_positions}
\alias{extract_filtered_igs_indel_positions}
\title{Extract Filtered IGS INDEL Positions}
\usage{
extract_filtered_igs_indel_positions(filtered_data_file)
}
\arguments{
\item{filtered_data_file}{Path to variants_filtered.csv}
}
\value{
Data frame with INDEL positions
}
\description{
Extract Filtered IGS INDEL Positions
}
\keyword{internal}
