% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M01_helpers_relational.R
\name{extract_flanking_genes_from_poigs}
\alias{extract_flanking_genes_from_poigs}
\title{Extract Flanking Genes from poiGS Identifier}
\usage{
extract_flanking_genes_from_poigs(poigs_id)
}
\arguments{
\item{poigs_id}{Character vector of poiGS identifiers}
}
\value{
Data frame with gene1, gene2, gene_pair
}
\description{
Extract Flanking Genes from poiGS Identifier
}
\keyword{internal}
