% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_processing.R
\name{extract_genome_structure}
\alias{extract_genome_structure}
\title{Extract genome structure information (LSC/IR/SSC regions)}
\usage{
extract_genome_structure(gff_object, species_name)
}
\arguments{
\item{gff_object}{Validated GRanges object}

\item{species_name}{Species name}
}
\value{
Data frame with genome structure information
}
\description{
Extract genome structure information (LSC/IR/SSC regions)
}
