% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M01_helpers_sequence_variants.R
\name{extract_indel_length_info}
\alias{extract_indel_length_info}
\title{Extract INDEL Length Information}
\usage{
extract_indel_length_info(preprocessed_data_file, whitelist)
}
\arguments{
\item{preprocessed_data_file}{Path to P01 data}

\item{whitelist}{INDEL position whitelist}
}
\value{
Data frame with INDEL details
}
\description{
Extract INDEL Length Information
}
\keyword{internal}
