% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M01_helpers_sequence_variants.R
\name{extract_sample_level_annotations}
\alias{extract_sample_level_annotations}
\title{Extract Sample Level Annotations}
\usage{
extract_sample_level_annotations(preprocessed_data_file, whitelist)
}
\arguments{
\item{preprocessed_data_file}{Path to P01 preprocessed data}

\item{whitelist}{Data frame with species and position}
}
\value{
Data frame with sample annotations
}
\description{
Extract Sample Level Annotations
}
\keyword{internal}
