% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_final_report.R
\name{format_duration}
\alias{format_duration}
\title{Format duration in human-readable format}
\usage{
format_duration(duration)
}
\arguments{
\item{duration}{difftime object}
}
\value{
Formatted duration string
}
\description{
Format duration in human-readable format
}
