% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_final_report.R
\name{format_parameters_for_display}
\alias{format_parameters_for_display}
\title{Format analysis parameters for display in reports (V7.1 Enhancement)}
\usage{
format_parameters_for_display(parameters)
}
\arguments{
\item{parameters}{task_params object or other parameters object}
}
\value{
Character string containing formatted Markdown parameters
}
\description{
Format analysis parameters for display in reports (V7.1 Enhancement)
}
