% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_distribution_histogram.R
\name{generate_distribution_histogram}
\alias{generate_distribution_histogram}
\title{Generate distribution histogram}
\usage{
generate_distribution_histogram(data, config, task_params)
}
\arguments{
\item{data}{Data frame with frequency data}

\item{config}{Configuration object}

\item{task_params}{Task parameters}
}
\value{
List containing plot, data, and metadata
}
\description{
Generate distribution histogram
}
