% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_dispatcher.R
\name{generate_execution_summary}
\alias{generate_execution_summary}
\title{Generate execution summary}
\usage{
generate_execution_summary(execution_results, parsed_config)
}
\arguments{
\item{execution_results}{Execution results}

\item{parsed_config}{Parsed configuration}
}
\value{
Execution summary
}
\description{
Generate execution summary
}
