% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_module_utils.R
\name{generate_filename}
\alias{generate_filename}
\title{Generate standardized filename}
\usage{
generate_filename(module_name, plot_type, factors = NULL, extension = "png")
}
\arguments{
\item{module_name}{Module name (e.g. "M01", "M02", "M03")}

\item{plot_type}{Plot type (e.g. "boxplot", "heatmap", "density")}

\item{factors}{Vector of factor names}

\item{extension}{File extension}
}
\value{
Standardized filename
}
\description{
Generate standardized filename
}
