% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_M01_sequence_variant_plots.R
\name{generate_genome_region_snp_indel_ratio_plot}
\alias{generate_genome_region_snp_indel_ratio_plot}
\title{Generate SNP/INDEL Ratio Plot by Genome Region}
\usage{
generate_genome_region_snp_indel_ratio_plot(region_stats, output_file, config)
}
\arguments{
\item{region_stats}{Data frame with columns: species, genome_region, snp, INDEL, snp_indel_ratio}

\item{output_file}{Path to save the plot}

\item{config}{Configuration object}
}
\description{
Create grouped barplot showing SNP/INDEL ratio across genome regions
}
\keyword{internal}
