% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_M01_sequence_variant_plots.R
\name{generate_indel_length_histogram}
\alias{generate_indel_length_histogram}
\title{Generate INDEL Length Histogram}
\usage{
generate_indel_length_histogram(
  indel_data,
  output_file,
  config,
  min_indel_length
)
}
\description{
Generate INDEL Length Histogram
}
\keyword{internal}
