% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_M01_sequence_variant_plots.R
\name{generate_indel_summary_comparison_plot}
\alias{generate_indel_summary_comparison_plot}
\title{Generate INDEL Summary Comparison Plot}
\usage{
generate_indel_summary_comparison_plot(
  species_stats_filtered,
  output_file,
  config,
  min_indel_length
)
}
\description{
Generate INDEL Summary Comparison Plot
}
\keyword{internal}
