% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pdf_reports.R
\name{generate_pdf_reports}
\alias{generate_pdf_reports}
\title{Generate PDF reports placeholder}
\usage{
generate_pdf_reports(data = NULL)
}
\arguments{
\item{data}{Input data}
}
\description{
Generate PDF reports placeholder
}
