% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_validate_region_info_v2.R
\name{generate_quality_report_v2}
\alias{generate_quality_report_v2}
\title{Generate Quality Report V2}
\usage{
generate_quality_report_v2(all_issues, output_path)
}
\description{
Generate Quality Report V2
}
\keyword{internal}
