% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_validator.R
\name{generate_recommendations}
\alias{generate_recommendations}
\title{Generate recommendations based on validation results}
\usage{
generate_recommendations(config_df, validation_result)
}
\arguments{
\item{config_df}{Configuration data frame}

\item{validation_result}{Validation results}
}
\value{
Character vector of recommendations
}
\description{
Generate recommendations based on validation results
}
