% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_significance_stars.R
\name{generate_significance_stars}
\alias{generate_significance_stars}
\title{Generate significance star notation}
\usage{
generate_significance_stars(
  pvalue_matrix,
  significance_levels = list(`***` = 0.001, `**` = 0.01, `*` = 0.05, . = 0.1, ns = 1),
  show_nonsignificant = FALSE
)
}
\arguments{
\item{pvalue_matrix}{Matrix of p-values to convert}

\item{significance_levels}{Named list defining star levels and thresholds}

\item{show_nonsignificant}{Whether to show non-significant values (default: FALSE)}
}
\value{
Matrix of significance stars
}
\description{
Converts p-value matrix to significance star notation
}
\details{
Generate significance star notation
}
