% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_M01_sequence_variant_plots.R
\name{generate_site_type_composition_barplot}
\alias{generate_site_type_composition_barplot}
\title{Generate Site Type Composition Barplot}
\usage{
generate_site_type_composition_barplot(
  site_classification,
  output_file,
  config
)
}
\description{
Generate Site Type Composition Barplot
}
\keyword{internal}
