% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_mapping.R
\name{generate_smart_mapping}
\alias{generate_smart_mapping}
\title{Generate smart column mapping suggestions with file type support}
\usage{
generate_smart_mapping(column_names, file_type = "main_data")
}
\arguments{
\item{column_names}{Vector of column names from user's data}

\item{file_type}{Type of file ("main_data", "cds_lengths", "region_info", "genome_regions", "group_info")}
}
\value{
List of mapping suggestions for each expected column
}
\description{
Generate smart column mapping suggestions with file type support
}
