% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_M01_sequence_variant_plots.R
\name{generate_sn_ratio_barplot}
\alias{generate_sn_ratio_barplot}
\title{Generate S/N Ratio Barplot}
\usage{
generate_sn_ratio_barplot(species_stats, output_file, config)
}
\description{
Generate S/N Ratio Barplot
}
\keyword{internal}
