% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_M01_sequence_variant_plots.R
\name{generate_snp_indel_count_barplot}
\alias{generate_snp_indel_count_barplot}
\title{Generate SNP/INDEL Count Barplot}
\usage{
generate_snp_indel_count_barplot(species_stats, output_file, config)
}
\description{
Generate SNP/INDEL Count Barplot
}
\keyword{internal}
