% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_M01_sequence_variant_plots.R
\name{generate_species_clustering_dendrogram}
\alias{generate_species_clustering_dendrogram}
\title{Generate Species Clustering Dendrogram}
\usage{
generate_species_clustering_dendrogram(hc, output_file, config)
}
\description{
Generate Species Clustering Dendrogram
}
\keyword{internal}
