% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_M01_sequence_variant_plots.R
\name{generate_syn_vs_nonsyn_scatter}
\alias{generate_syn_vs_nonsyn_scatter}
\title{Generate Syn vs Nonsyn Scatter}
\usage{
generate_syn_vs_nonsyn_scatter(species_stats, output_file, config)
}
\description{
Generate Syn vs Nonsyn Scatter
}
\keyword{internal}
