% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_validate_region_info_v2.R
\name{generate_txt_report_v2}
\alias{generate_txt_report_v2}
\title{Generate TXT Report V2}
\usage{
generate_txt_report_v2(report, output_path)
}
\description{
Generate TXT Report V2
}
\keyword{internal}
