% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_validator.R
\name{generate_validation_report}
\alias{generate_validation_report}
\title{Generate detailed validation report}
\usage{
generate_validation_report(config_df, schema, validation_result)
}
\arguments{
\item{config_df}{Configuration data frame}

\item{schema}{Configuration schema}

\item{validation_result}{Validation results}
}
\value{
Detailed report list
}
\description{
Generate detailed validation report
}
