% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_alpha_scheme.R
\name{get_alpha_scheme}
\alias{get_alpha_scheme}
\title{Get alpha transparency scheme}
\usage{
get_alpha_scheme(scheme_name = "standard")
}
\arguments{
\item{scheme_name}{Name of the alpha scheme (default: "standard")}
}
\value{
List of alpha values for different plot elements
}
\description{
Returns predefined alpha transparency values for plot elements
}
\details{
Get alpha transparency scheme
}
