% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_application_theme.R
\name{get_application_theme}
\alias{get_application_theme}
\title{Get application theme}
\usage{
get_application_theme(
  theme_name = "professional_light",
  custom_theme_file = NULL,
  base_size = 11,
  base_family = "Arial"
)
}
\arguments{
\item{theme_name}{Name of the theme to apply}

\item{custom_theme_file}{Path to custom theme file}

\item{base_size}{Base font size}

\item{base_family}{Base font family}
}
\value{
ggplot2 theme object
}
\description{
Get application theme
}
