% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_module_pdfs.R
\name{get_available_pdf_reports}
\alias{get_available_pdf_reports}
\title{Get list of available PDF reports for a session}
\usage{
get_available_pdf_reports(session_path)
}
\arguments{
\item{session_path}{Character. Path to the session directory}
}
\value{
Character vector. Paths to available PDF reports
}
\description{
Get list of available PDF reports for a session
}
