% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_parser.R
\name{get_execution_plan_summary}
\alias{get_execution_plan_summary}
\title{Get execution plan summary}
\usage{
get_execution_plan_summary(execution_plan)
}
\arguments{
\item{execution_plan}{Task execution plan}
}
\value{
Human-readable execution summary
}
\description{
Get execution plan summary
}
