% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_factor_levels.R
\name{get_factor_levels}
\alias{get_factor_levels}
\title{Get ordered factor levels}
\usage{
get_factor_levels(factor_values, factor_name, preferred_order = NULL)
}
\arguments{
\item{factor_values}{Vector of factor values to order}

\item{factor_name}{Name of the factor being processed}

\item{preferred_order}{Preferred ordering of factor levels (optional)}
}
\value{
Ordered vector of factor levels
}
\description{
Determines factor level ordering based on preferences or data characteristics
}
\details{
Get ordered factor levels
}
