% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_mapping.R
\name{get_field_descriptions}
\alias{get_field_descriptions}
\title{Get field descriptions for interactive help}
\usage{
get_field_descriptions(file_type = "main_data")
}
\arguments{
\item{file_type}{Type of file}
}
\value{
List of field descriptions
get_field_descriptions
}
\description{
Get field descriptions for interactive help
}
