% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ggsci_colors.R
\name{get_ggsci_colors}
\alias{get_ggsci_colors}
\title{Get scientific color palette from ggsci}
\usage{
get_ggsci_colors(palette, n_colors = NULL)
}
\arguments{
\item{palette}{Name of the ggsci palette to use}

\item{n_colors}{Number of colors to retrieve (optional)}
}
\value{
Vector of color codes
}
\description{
Retrieves color palettes from the ggsci package for scientific publications
}
\details{
Get scientific color palette from ggsci
}
