% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_helpers.R
\name{get_heatmap_colors}
\alias{get_heatmap_colors}
\title{Create heatmap color gradient}
\usage{
get_heatmap_colors(type = "hotspot")
}
\arguments{
\item{type}{Type of heatmap colors to generate}
}
\value{
Vector of color values
}
\description{
Create heatmap color gradient
}
