% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M01_helpers_chromosome_hotspot.R
\name{get_hotspot_colors_from_config}
\alias{get_hotspot_colors_from_config}
\title{Get Hotspot Colors from Configuration}
\usage{
get_hotspot_colors_from_config(config)
}
\arguments{
\item{config}{Configuration object}
}
\value{
Named vector with 'cds' and 'poigs' colors
}
\description{
Retrieves color palette for CDS and poiGS hotspots from config
}
\keyword{internal}
