% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_module_utils.R
\name{get_main_theme}
\alias{get_main_theme}
\title{Get unified ggplot theme}
\usage{
get_main_theme(base_size = 12, title_size = 14)
}
\arguments{
\item{base_size}{Base font size}

\item{title_size}{Title font size}
}
\value{
ggplot theme object
}
\description{
Get unified ggplot theme
}
