% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_helpers.R
\name{get_region_type_colors}
\alias{get_region_type_colors}
\title{Get standardized color palette for region types}
\usage{
get_region_type_colors()
}
\description{
Get standardized color palette for region types
}
