% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_task_parameter.R
\name{get_task_parameter}
\alias{get_task_parameter}
\title{Get task parameter value with hierarchical lookup}
\usage{
get_task_parameter(
  task_params,
  config,
  param_name,
  default = NULL,
  log_default = FALSE
)
}
\arguments{
\item{task_params}{Task parameters}

\item{config}{Configuration object}

\item{param_name}{Parameter name}

\item{default}{Default value if not found}

\item{log_default}{Whether to log when using default}
}
\value{
Parameter value
}
\description{
Get task parameter value with hierarchical lookup
}
