% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M01_helpers_sequence_variants.R
\name{identify_shared_indel_sequences}
\alias{identify_shared_indel_sequences}
\title{Identify Shared INDEL Sequences}
\usage{
identify_shared_indel_sequences(indel_data, min_species)
}
\arguments{
\item{indel_data}{Sample-level INDEL data}

\item{min_species}{Minimum species threshold}
}
\value{
Data frame with shared sequences
}
\description{
Identify Shared INDEL Sequences
}
\keyword{internal}
