% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_chromosome_distribution_plot.R
\name{lighten_color}
\alias{lighten_color}
\title{Lighten a hex color by a specified amount}
\usage{
lighten_color(hex_color, amount = 0.5)
}
\arguments{
\item{hex_color}{A hex color string (e.g., "#3B4992FF")}

\item{amount}{Lightening factor (0-1, where 1 is white)}
}
\value{
Lightened hex color
}
\description{
Lighten a hex color by a specified amount
}
\keyword{internal}
