% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_frequencies.R
\name{load_gene_lengths}
\alias{load_gene_lengths}
\title{Load gene lengths data}
\usage{
load_gene_lengths(gene_lengths_file)
}
\arguments{
\item{gene_lengths_file}{Path to gene lengths file}
}
\value{
Data frame with gene lengths
}
\description{
Load gene lengths data
}
