% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_frequencies.R
\name{load_region_info}
\alias{load_region_info}
\title{Load region information data (intergenic and intron regions)}
\usage{
load_region_info(region_info_file, config)
}
\arguments{
\item{region_info_file}{Path to region info file}

\item{config}{Configuration list}
}
\value{
Data frame with region information including coordinates
}
\description{
Load region information data (intergenic and intron regions)
}
