% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_dispatcher.R
\name{merge_parameters}
\alias{merge_parameters}
\title{Merge global and task-specific parameters}
\usage{
merge_parameters(global_params, task_params)
}
\arguments{
\item{global_params}{Global parameters}

\item{task_params}{Task-specific parameters}
}
\value{
Merged parameters
}
\description{
Merge global and task-specific parameters
}
