% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M01_helpers_relational.R
\name{merge_poigs_gene_frequencies}
\alias{merge_poigs_gene_frequencies}
\title{Merge poiGS and Gene Frequency Data}
\usage{
merge_poigs_gene_frequencies(
  poigs_freq,
  gene_freq,
  poigs_mapping,
  keep_all_zeros = TRUE
)
}
\arguments{
\item{poigs_freq}{Data frame with poiGS frequency data}

\item{gene_freq}{Data frame with gene frequency data}

\item{poigs_mapping}{Data frame with poiGS mapping}

\item{keep_all_zeros}{Whether to keep pairs where both values are zero.}
}
\value{
Merged data frame
}
\description{
Links each poiGS with average frequency of its flanking genes
}
\keyword{internal}
