% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_helpers.R
\name{mytheme}
\alias{mytheme}
\title{Professional scientific publication theme (mytheme)}
\usage{
mytheme(x_angle = 0, legend_position = "right")
}
\arguments{
\item{x_angle}{Angle for x-axis text labels}

\item{legend_position}{Position of the legend}
}
\value{
ggplot2 theme object
}
\description{
Professional scientific publication theme (mytheme)
}
