% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_processing.R
\name{null.coalesce}
\alias{null.coalesce}
\alias{\%||\%}
\title{Null coalescing operator}
\usage{
x \%||\% y
}
\arguments{
\item{x}{First value to test (left-hand side)}

\item{y}{Alternative value if \code{x} is NULL}
}
\value{
Either \code{x} (if not NULL) or \code{y}
}
\description{
Returns the left-hand side if not NULL, otherwise returns the right-hand side.
This is a standard null-coalescing operator commonly used in R packages.
}
\examples{
# Basic usage
NULL \%||\% "default"        # Returns "default"
"value" \%||\% "default"     # Returns "value"

# Common use case
user_input <- NULL
final_value <- user_input \%||\% "default_value"

}
\keyword{internal}
