% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_parser.R
\name{parse_analysis_tasks}
\alias{parse_analysis_tasks}
\title{Parse individual analysis tasks}
\usage{
parse_analysis_tasks(tasks_config)
}
\arguments{
\item{tasks_config}{Analysis tasks configuration}
}
\value{
List containing parsed tasks and metadata
}
\description{
Parse individual analysis tasks
}
