% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_parser.R
\name{parse_single_task}
\alias{parse_single_task}
\title{Parse a single task configuration}
\usage{
parse_single_task(task_name, task_config)
}
\arguments{
\item{task_name}{Name of the task}

\item{task_config}{Task configuration}
}
\value{
Parsed task object
}
\description{
Parse a single task configuration
}
