% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M01_helpers_sequence_variants.R
\name{parse_variant_type_from_v12}
\alias{parse_variant_type_from_v12}
\title{Parse Variant Type}
\usage{
parse_variant_type_from_v12(sample_annotations)
}
\arguments{
\item{sample_annotations}{Data frame with functional_annotation}
}
\value{
Data frame with variant_class
}
\description{
Parse Variant Type
}
\keyword{internal}
