% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_statistical_comparisons.R
\name{perform_statistical_comparisons}
\alias{perform_statistical_comparisons}
\title{Perform statistical comparisons between groups}
\usage{
perform_statistical_comparisons(data, group_col, value_col)
}
\arguments{
\item{data}{Dataset containing group and value columns}

\item{group_col}{Name of the grouping column}

\item{value_col}{Name of the value column to compare}
}
\value{
List containing statistical test results
}
\description{
Conducts statistical tests to compare values between different groups
}
\details{
Perform statistical comparisons between groups
}
