% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{pivot_to_long}
\alias{pivot_to_long}
\title{Convert data to long format for analysis
Pivot data to long format}
\usage{
pivot_to_long(
  data,
  id_cols,
  measure_cols,
  measure_name = "variable",
  value_name = "value"
)
}
\arguments{
\item{data}{Data frame to pivot}

\item{id_cols}{ID columns}

\item{measure_cols}{Measure columns}

\item{measure_name}{Name for measure column}

\item{value_name}{Name for value column}
}
\value{
Long format data frame
}
\description{
Convert data to long format for analysis
Pivot data to long format
}
