% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_chromosome_hotspot_ideogram.R
\name{plot_hotspot_ideogram}
\alias{plot_hotspot_ideogram}
\title{Plot Hotspot Ideogram}
\usage{
plot_hotspot_ideogram(
  karyotype_data,
  regions_heatmap,
  hotspot_markers,
  output_dir,
  config
)
}
\arguments{
\item{karyotype_data}{Karyotype data frame}

\item{regions_heatmap}{CDS and IGS regions heatmap data}

\item{hotspot_markers}{Hotspot marker data frame}

\item{output_dir}{Output directory}

\item{config}{Configuration object}
}
\value{
Path to generated plot
}
\description{
Generates the actual ideogram plot using RIdeogram with overlaid heatmap
}
\keyword{internal}
