% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M01_helpers_chromosome_hotspot.R
\name{prepare_regions_heatmap}
\alias{prepare_regions_heatmap}
\title{Prepare CDS and IGS Regions Heatmap Data}
\usage{
prepare_regions_heatmap(region_info_path, genome_regions)
}
\arguments{
\item{region_info_path}{Path to region_info_complete.csv}

\item{genome_regions}{Genome regions data with Chr mapping}
}
\value{
Data frame with columns: Chr, Start, End, Value
}
\description{
Extracts all CDS and IGS regions from region_info_complete.csv
and prepares them for RIdeogram overlaid heatmap visualization.
Uses a diverging value scale: CDS = 1 (red), IGS = -1 (blue), dummy = 0 (white)
}
\keyword{internal}
