% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M01_helpers_sequence_variants.R
\name{prepare_upset_matrix_by_sequence}
\alias{prepare_upset_matrix_by_sequence}
\title{Prepare UpSet Matrix by Sequence}
\usage{
prepare_upset_matrix_by_sequence(indel_data)
}
\arguments{
\item{indel_data}{Sample-level INDEL data}
}
\value{
Binary matrix for UpSet plot
}
\description{
Prepare UpSet Matrix by Sequence
}
\keyword{internal}
