% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_dispatcher.R
\name{route_task_to_module}
\alias{route_task_to_module}
\title{Route task to appropriate module function}
\usage{
route_task_to_module(task_config, execution_context, progress_callback = NULL)
}
\arguments{
\item{task_config}{Task configuration}

\item{execution_context}{Execution context}

\item{progress_callback}{Progress callback function}
}
\value{
Module execution result
}
\description{
Route task to appropriate module function
}
