% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_final_report.R
\name{safe_sprintf}
\alias{safe_sprintf}
\title{Safe sprintf wrapper to prevent list coercion errors}
\usage{
safe_sprintf(fmt, ...)
}
\arguments{
\item{fmt}{Format string}

\item{...}{Arguments to be formatted}
}
\value{
Formatted string
}
\description{
Safe sprintf wrapper to prevent list coercion errors
}
