% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_yaml_utils.R
\name{shiny_to_yaml_config}
\alias{shiny_to_yaml_config}
\title{Convert Shiny reactive values to YAML configuration}
\usage{
shiny_to_yaml_config(analysis_data, input, session_paths)
}
\arguments{
\item{analysis_data}{Shiny reactiveValues object containing user selections}

\item{input}{Shiny input object containing UI parameters}

\item{session_paths}{Session directory paths}
}
\value{
Configuration list ready for YAML export
}
\description{
Convert Shiny reactive values to YAML configuration
}
