% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_variants.R
\name{standardize_column_names}
\alias{standardize_column_names}
\title{Standardize column names in raw data}
\usage{
standardize_column_names(data)
}
\arguments{
\item{data}{Raw variant data (either with V columns or already mapped columns)}
}
\value{
Data with standardized column names
}
\description{
Standardize column names in raw data
}
