% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{standardize_variant_types}
\alias{standardize_variant_types}
\title{Standardize variant type names}
\usage{
standardize_variant_types(var_types)
}
\arguments{
\item{var_types}{Character vector of variant types to standardize}
}
\description{
Standardize variant type names
}
