% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_validator.R
\name{validate_business_rules}
\alias{validate_business_rules}
\title{Validate business rules}
\usage{
validate_business_rules(config_df, rules_schema)
}
\arguments{
\item{config_df}{Configuration data frame}

\item{rules_schema}{Business rules from schema}
}
\value{
Validation result list
}
\description{
Validate business rules
}
