% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_mapping.R
\name{validate_column_mapping}
\alias{validate_column_mapping}
\title{Validate column mapping completeness with file type support}
\usage{
validate_column_mapping(mapping_list, file_type = "main_data")
}
\arguments{
\item{mapping_list}{List of user-selected column mappings}

\item{file_type}{Type of file}
}
\value{
List with validation results
}
\description{
Validate column mapping completeness with file type support
}
