% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_validator.R
\name{validate_columns}
\alias{validate_columns}
\title{Validate column structure against schema}
\usage{
validate_columns(config_df, column_schema)
}
\arguments{
\item{config_df}{Configuration data frame}

\item{column_schema}{Schema column definitions}
}
\value{
Validation result list
}
\description{
Validate column structure against schema
}
