% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_parser.R
\name{validate_input_files}
\alias{validate_input_files}
\title{Validate input files configuration}
\usage{
validate_input_files(input_files_config)
}
\arguments{
\item{input_files_config}{Input files configuration}
}
\value{
File validation results
}
\description{
Validate input files configuration
}
