% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_module_utils.R
\name{validate_required_columns}
\alias{validate_required_columns}
\title{Validate required columns in data frame}
\usage{
validate_required_columns(data, required_cols)
}
\arguments{
\item{data}{Data frame}

\item{required_cols}{Vector of required column names}
}
\value{
Logical value, TRUE if all required columns exist
}
\description{
Validate required columns in data frame
}
