% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loader.R
\name{validate_standardized_data}
\alias{validate_standardized_data}
\title{Validate standardized data quality}
\usage{
validate_standardized_data(data, file_description = "data")
}
\arguments{
\item{data}{Standardized data frame}

\item{file_description}{Description for messages}
}
\value{
List with valid (boolean) and message (string)
}
\description{
Validate standardized data quality
}
