% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_parser.R
\name{validate_task_dependencies}
\alias{validate_task_dependencies}
\title{Validate task dependencies}
\usage{
validate_task_dependencies(tasks)
}
\arguments{
\item{tasks}{List of parsed tasks}
}
\value{
Dependency validation results
}
\description{
Validate task dependencies
}
