% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_parser.R
\name{validate_task_parameters}
\alias{validate_task_parameters}
\title{Validate task parameters based on module type}
\usage{
validate_task_parameters(task)
}
\arguments{
\item{task}{Parsed task object}
}
\value{
Parameter validation results
}
\description{
Validate task parameters based on module type
}
