% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_yaml_utils.R
\name{validate_yaml_config}
\alias{validate_yaml_config}
\title{Validate YAML configuration file}
\usage{
validate_yaml_config(config_file, return_errors = FALSE)
}
\arguments{
\item{config_file}{Path to YAML configuration file}

\item{return_errors}{If TRUE, return detailed validation errors}
}
\value{
TRUE if valid, FALSE if invalid, or list of errors if return_errors=TRUE
validate_yaml_config
}
\description{
Validate YAML configuration file
}
