% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_chromosome_hotspot_ideogram.R
\name{write_hotspot_summary}
\alias{write_hotspot_summary}
\title{Write Hotspot Summary Statistics}
\usage{
write_hotspot_summary(
  cds_hotspots,
  poigs_hotspots,
  hotspot_markers,
  output_path
)
}
\arguments{
\item{cds_hotspots}{CDS hotspot data}

\item{poigs_hotspots}{poiGS hotspot data}

\item{hotspot_markers}{Combined marker data}

\item{output_path}{Output file path}
}
\description{
Generates a text summary of hotspot statistics
}
\keyword{internal}
